fixing bugs and add features

+++ src/compression.py
@@ -1,5 +1,3 @@
 import zlib
-
 def compress_data(data):
-    return zlib.compress(data.tobytes())
-#test456
+    return zlib.compress(data.astype(np.float32).tobytes())
